# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "summary": "Get current authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "summary": "Update current authenticated user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/users/me/password": {
      "post": {
        "operationId": "UsersController_changePassword",
        "summary": "Change password for the current authenticated user",
        "description": "Requires the current password for verification. The current JWT remains valid after a password change.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "example": {
                        "message": "Password updated successfully"
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "New password is the same as current, or fails validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Current password is incorrect",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/words": {
      "post": {
        "operationId": "WordsController_create",
        "summary": "Create a new word in the user vocabulary",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Word created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WordResponseDto"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Words"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "get": {
        "operationId": "WordsController_findAll",
        "summary": "List words owned by the current user (paginated)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (capped at 100)",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "timestamp"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "items",
                        "total",
                        "page",
                        "limit",
                        "hasMore"
                      ],
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WordResponseDto"
                          }
                        },
                        "total": {
                          "type": "integer",
                          "example": 137
                        },
                        "page": {
                          "type": "integer",
                          "example": 1
                        },
                        "limit": {
                          "type": "integer",
                          "example": 20
                        },
                        "hasMore": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Words"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/words/{id}": {
      "delete": {
        "operationId": "WordsController_remove",
        "summary": "Delete a word owned by the current user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Word deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "example": {
                        "message": "Word deleted successfully"
                      }
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Word belongs to another user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Word not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Words"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EnglishFlow API",
    "description": "REST API for the EnglishFlow vocabulary learning app. All successful responses are wrapped in `{ success, data, timestamp }` and all errors in `{ success: false, statusCode, message, error, errors?, path, timestamp }`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "8c5f4e64-1c39-4a55-9c8e-1e1b9c41a0a2",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "student@example.com",
            "format": "email"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-11T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-11T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "errors": {
            "example": [
              "email must be an email"
            ],
            "description": "Present for validation errors",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-11T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "error",
          "path",
          "timestamp"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "new-email@example.com",
            "description": "New email address. Must be unique."
          }
        }
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "CurrentPass123!",
            "description": "The user's current password"
          },
          "newPassword": {
            "type": "string",
            "example": "NewStrongerPass456!",
            "minLength": 6,
            "description": "The new password to set"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "WordResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "translation": {
            "type": "string",
            "example": "kutilmagan yoqimli kashfiyot"
          },
          "example": {
            "type": "object",
            "example": "Finding that book was pure serendipity."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "word",
          "translation",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateWordDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "example": "serendipity"
          },
          "translation": {
            "type": "string",
            "example": "kutilmagan yoqimli kashfiyot"
          },
          "example": {
            "type": "string",
            "example": "Finding that book was pure serendipity."
          }
        },
        "required": [
          "word",
          "translation"
        ]
      },
...
    }
  }
}